# BUG: EWM.__getitem__ raised error with times