- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing.html#code-standards) for how to run them
- [x] whatsnew entry


Also somewhat ameliorates the performance hit in https://github.com/pandas-dev/pandas/pull/40072#issuecomment-787456038

```
dtype="float"
window=1000
N = 10 ** 5
arr = (100 * np.random.random(N)).astype(dtype)
ewm = pd.Series(arr).ewm(halflife=window)

%timeit ewm.mean()
# Master
2.59 ms ± 25.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# PR
2.5 ms ± 31.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```