@jreback 2021-03-02
does this address https://github.com/pandas-dev/pandas/issues/40099?