@mroeschke 2021-03-03
> does this address #40099?

No, that issue has the expected output given the formula we're using but needs a discussion whether we want to make an exception in that case